13 Ai agent 6/requirements.txt
# Multi-Modal AI Agent Requirements

# Core LangChain and LangGraph dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0

# OpenAI API
openai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Data processing and utilities
pandas>=1.5.0
numpy>=1.24.0

# File operations and utilities
pathlib2>=2.3.0

# Web requests (for web search functionality)
requests>=2.28.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# System monitoring (optional)
psutil>=5.9.0

# JSON and data handling
jsonschema>=4.17.0

# Text processing
nltk>=3.8.0
textblob>=0.17.0

# Image processing (optional)
Pillow>=9.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
